Add custom widget for dashboard bottom nav bar
